+++ src/main/java/org/fundacionjala/pivotal/model/pageobjects/WebDriverManager.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.pivotal.model.pageObject;
+package org.fundacionjala.pivotal.model.pageobjects;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
